Changes to implement a Voice Activity Detector in the RTP stream.
This effectively implements silence suppression in the outgoing RTP stream
but stopping the transmission of packets when silence is detected.

Two news configuration options have been added VADOn (which controls if the behaviour is on or not)
and VADmsBeforeSuppression which controls the number of milliseconds of silence that is allowed before
we stop sending the RTP packets.

   ~benm